requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-filter==22.1
django-rest==0.8.7
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
zipp==3.9.0
